commits-plotter/requirements.txt
GitPython==3.1.43
matplotlib==3.7.5
drawilleplot==0.1.0
